Move ReaderWebView to its own package
